requirements.txt
Django==5.2.6
psycopg2-binary==2.9.10
asgiref==3.9.1
sqlparse==0.5.3
typing-extensions==4.15.0
Pillow==10.0.0
stripe==8.0.0
python-dotenv==1.0.0
twilio==9.0.0
django-q2==1.8.0
requests==2.31.0